Enabled memory tracking with callstacks for PS4/5
Implementation details:
- PS4 & PS5 uses dwarf symbol format and uses Syms symbol resolver.
- To resolve symbols the path to folder where .self file is built currently must be specified in UE_INSIGHTS_SYMBOL_PATH env variable for Insights.
- Multiple paths can be separated by ; in this variable.
- Build for PS5 does not seem to have PLATFORM_PS5 define, I used defined(__PS5__).
- PS5 runtime collects and sends callstacks, but Syms resolver does not support dwarf v5 format yet, which is used on PS5 toolchain.

#rb none
#preflight 6112ead49c7bb10001bc63fc

[CL 17128247 by martins mozeiko in ue5-main branch]

+++ Engine/Source/Runtime/Core/Public/Experimental/Containers/SherwoodHashTable.h
@@ -226,7 +226,7 @@ struct TSherwoodHashTable
 
 	FORCENOINLINE ValueType* Add(int8 Distance, FData& Data, uint32 Cursor, KeyType Key, HashType Hash, ValueType Value)
 	{
-		if (Distance == MaxLookups || NumElements + 1 > (NumSlotsMinusOne + 1) * MaxLoadFactor)
+		if (Distance == MaxLookups || NumElements + 1 > int32(float(NumSlotsMinusOne + 1) * MaxLoadFactor))
 		{
 			Grow();
 			return FindOrAddByHash(Key, Hash, Value);
@@ -286,7 +286,7 @@ struct TSherwoodHashTable
 
 	void Rehash(uint32 DesiredNumSlots)
 	{
-		DesiredNumSlots = FMath::Max(DesiredNumSlots, uint32(FMath::CeilToInt(NumElements / MaxLoadFactor)));
+		DesiredNumSlots = FMath::Max(DesiredNumSlots, uint32(FMath::CeilToInt(float(NumElements) / MaxLoadFactor)));
 		if (DesiredNumSlots == 0)
 		{
 			Empty();
@@ -338,7 +338,7 @@ struct TSherwoodHashTable
 
 	void Reserve(uint32 DesiredNumElements)
 	{
-		uint32 DesiredNumSlots = FMath::Max<uint32>(DesiredNumElements, uint32(FMath::CeilToInt(DesiredNumElements / MaxLoadFactor)));
+		uint32 DesiredNumSlots = FMath::Max<uint32>(DesiredNumElements, uint32(FMath::CeilToInt(float(DesiredNumElements) / MaxLoadFactor)));
 		if (DesiredNumSlots > NumSlots())
 		{
 			Rehash(DesiredNumSlots);
